app/requirements.txt
Flask==2.3.3
weaviate-client==3.24.1
torchvision==0.15.2